Move state classes into their own file
