+++ src/state/firehose.js
@@ -1,0 +1,3 @@
+class Firehose {}
+
+module.exports = Firehose
