+++ src/state/globalState.js
@@ -1,25 +1,7 @@
-/*
- * Note: S3, Lambda, etc classes will be moved into their own file soon.
- */
-
-// eslint-disable-next-line max-classes-per-file
-class S3 {}
-
-class Lambda {
-  constructor({
-    Role,
-    Policy,
-    FunctionData,
-  }) {
-    this.Role = Role
-    this.Policy = Policy
-    this.FunctionData = FunctionData
-  }
-}
-
-class Firehose {}
-
-class Timestream {}
+const Lambda = require('./lambda')
+const S3 = require('./s3')
+const Firehose = require('./firehose')
+const Timestream = require('./timestream')
 
 /*
  * GlobalState is an singleton that will contain all data over the lifetime
