+++ src/state/s3.js
@@ -1,0 +1,3 @@
+class S3 {}
+
+module.exports = S3
