+++ src/state/timestream.js
@@ -1,0 +1,3 @@
+class Timestream {}
+
+module.exports = Timestream
